Create README.md
My greatest wish is to soak up as much information as I can about cybersecurity and never stop learning. I want to delve into multiple areas of this fascinating subject and when the time is right, I hope to work for a large company if possible while giving 110% of my time and attention to every detail and project assigned.  I currently work two jobs that require 64 hours of my time and am committed to giving 30 hours of time to a University of Minnesota bootcamp. I will do the hard work to achieve my goals. My work ethic is very strong.